# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "User-Manager service is service for managing users.",
    "title": "User-Manager"
  },
  "paths": {
...
    "/login/basic": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#sign-in-with-login-and-password",
        "tags": [
          "Login"
        ],
        "summary": "Basic login.",
        "operationId": "BasicLoginHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserAgentHeader"
          },
          {
            "$ref": "#/parameters/FingerprintHeader"
          },
          {
            "$ref": "#/parameters/ClientIPHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user logged in",
            "schema": {
              "$ref": "#/definitions/CreateTokenResponse"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/login/oauth": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#sign-in-with-third-party-resource-oauth",
        "tags": [
          "Login"
        ],
        "summary": "Login using oauth service.",
        "operationId": "OAuthLoginRequest",
        "parameters": [
          {
            "$ref": "#/parameters/UserAgentHeader"
          },
          {
            "$ref": "#/parameters/FingerprintHeader"
          },
          {
            "$ref": "#/parameters/ClientIPHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/OAuthLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user logged in",
            "schema": {
              "$ref": "#/definitions/CreateTokenResponse"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/login/token": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#sign-in-with-one-time-token",
        "tags": [
          "Login"
        ],
        "summary": "Login with one-time token.",
        "operationId": "OneTimeTokenLoginHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserAgentHeader"
          },
          {
            "$ref": "#/parameters/FingerprintHeader"
          },
          {
            "$ref": "#/parameters/ClientIPHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/OneTimeTokenLoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user logged in",
            "schema": {
              "$ref": "#/definitions/CreateTokenResponse"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/logout": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#sign-out",
        "tags": [
          "Login"
        ],
        "summary": "Logout for users who used one-time token login.",
        "operationId": "LogoutHandler",
        "parameters": [
          {
            "$ref": "#/parameters/TokenIDHeader"
          },
          {
            "$ref": "#/parameters/SessionIDHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "user logged out"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/user/bound_accounts": {
      "get": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#get-bound-accounts",
        "tags": [
          "BoundAccounts"
        ],
        "summary": "Get users bound accounts.",
        "operationId": "GetBoundAccountsHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "bound accounts list",
            "schema": {
              "$ref": "#/definitions/BoundAccounts"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#add-bound-accounts",
        "tags": [
          "BoundAccounts"
        ],
        "summary": "Bind account.",
        "operationId": "GetBoundAccountsHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/OAuthLoginRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "account added"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#delete-bound-account",
        "tags": [
          "BoundAccounts"
        ],
        "summary": "Unbind account.",
        "operationId": "DeleteBoundAccountHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/BoundAccountDeleteRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "account deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/user/links/{user_id}": {
      "get": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#get-user-links",
        "tags": [
          "Links"
        ],
        "summary": "Get user links.",
        "operationId": "LinksGetHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "type": "string",
            "name": "user_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "link sent",
            "schema": {
              "$ref": "#/definitions/Links"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/user/sign_up/resend": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#resend-activation-link",
        "tags": [
          "Links"
        ],
        "summary": "Resend activation link.",
        "operationId": "LinkResendHandler",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "link sent"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    }
  },
  "definitions": {
...
    "BoundAccountDeleteRequest": {
      "description": "BoundAccountDeleteRequest -- request to remove bound account",
      "type": "object",
      "required": [
        "resource"
      ],
      "properties": {
        "resource": {
          "type": "string",
          "x-go-name": "Resource"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "BoundAccounts": {
      "description": "BoundAccounts -- bound accounts list for user",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "CreateTokenResponse": {
      "type": "object",
      "title": "CreateTokenResponse contains access and refresh token.",
      "properties": {
        "access_token": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "refresh_token": {
          "type": "string",
          "x-go-name": "RefreshToken"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/git.containerum.net/ch/auth/proto"
    },
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
...
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
...
    "Link": {
      "description": "Link -- link (for registration/activation/etc)",
      "type": "object",
      "required": [
        "link"
      ],
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "expired_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ExpiredAt"
        },
        "is_active": {
          "type": "boolean",
          "x-go-name": "IsActive"
        },
        "link": {
          "type": "string",
          "x-go-name": "Link"
        },
        "sent_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "SentAt"
        },
        "type": {
          "$ref": "#/definitions/LinkType"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "LinkType": {
      "description": "LinkType -- link type",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "Links": {
      "description": "Links -- links list",
      "type": "object",
      "properties": {
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Link"
          },
          "x-go-name": "Links"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
    "LoginRequest": {
      "description": "LoginRequest -- login request (for basic login)",
      "type": "object",
      "required": [
        "login",
        "password"
      ],
      "properties": {
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
    "OAuthLoginRequest": {
      "description": "LoginRequest -- login request (for oauth login)",
      "type": "object",
      "required": [
        "resource",
        "access_token"
      ],
      "properties": {
        "access_token": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "resource": {
          "$ref": "#/definitions/OAuthResource"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "OAuthResource": {
      "type": "string",
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "OneTimeTokenLoginRequest": {
      "description": "LoginRequest -- login request (for token login)",
      "type": "object",
      "required": [
        "token"
      ],
      "properties": {
        "token": {
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
    "UserLogin": {
      "description": "UserList -- model for user login, password and id",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
  },
  "parameters": {
    "ClientIPHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/ClientIPHeader"
    },
    "FingerprintHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserClientHeader"
    },
    "SessionIDHeader": {
      "type": "string",
      "name": "X-Session-ID",
      "in": "header",
      "required": true
    },
    "TokenIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/TokenIDHeader"
    },
    "UserAgentHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserAgentHeader"
    },
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}